+++ no_xlsx/DIDmultiplegtDYN/NEWS.md
@@ -27,3 +27,19 @@ First official R release.
 + Added option to compute bootstrap standard errors
 
 + Fixed minor bugs
+
+## Changes in version 1.0.6
+
++ Added by_path option.
+
+## Changes in version 1.0.7
+
++ Added average number of cumulative effects
+
++ Added compatibility with rnames to enhance results browsing experience
+
+## Changes in version 1.0.8
+
++ Changed display with weight option
+
++ Integration with honestdid via the "coef" output in the assigned object